+++ context/UserContext.js
@@ -4,9 +4,10 @@ const UserContext = createContext();
 
 export const UserProvider = ({ children }) => {
   const [user, setUser] = useState(null);
+  const [roomCode, setRoomCode] = useState(null);
 
   return (
-    <UserContext.Provider value={{ user, setUser }}>
+    <UserContext.Provider value={{ user, setUser, roomCode, setRoomCode }}>
       {children}
     </UserContext.Provider>
   );
