+++ package.json
@@ -6,10 +6,11 @@
     "start": "expo start",
     "android": "expo run:android",
     "ios": "expo run:ios",
-    "web": "expo start --web"
+    "web": "expo start --web",
+    "dev": "npx expo start --dev-client"
   },
   "dependencies": {
-    "@100mslive/react-native-hms": "^1.10.3",
+    "@100mslive/react-native-hms": "^1.10.4",
     "@expo/webpack-config": "~19.0.1",
     "@react-native-async-storage/async-storage": "1.21.0",
     "@react-native-google-signin/google-signin": "^11.0.0",
@@ -17,6 +18,7 @@
     "@react-navigation/native-stack": "^6.9.12",
     "@rneui/base": "^4.0.0-rc.7",
     "@rneui/themed": "^4.0.0-rc.7",
+    "axios": "^1.6.8",
     "expo": "^50.0.0",
     "expo-camera": "~14.1.3",
     "expo-dev-client": "~3.3.11",
